+++ utils/defaultsettings.gd
@@ -29,7 +29,13 @@ var overlay_rotation_x = 0
 var overlay_rotation_y = 0
 var overlay_rotation_z = 0
 
+var overlay_rotation_x_hand = 0
+var overlay_rotation_y_hand = 0
+var overlay_rotation_z_hand = 0
+
 # Overlay position / rotation (hand tracking)
+var overlay_position_x_hand = 0
+var overlay_position_y_hand = 0
 var overlay_position_z_hand = -0.54
 var overlay_rotation_z_hand_left = 1.7
 var overlay_rotation_z_hand_right = -1.7